+++ app/quantumChess/useQuantumCallapse.ts
@@ -1,24 +1,20 @@
-const calculateEntangledCollapse = () => {};
+// const calculateEntangledCollapse = () => {};
 
-const handleCollapse = (piece, triggerPosition) => {
-  if (piece.isSuperposed) {
-    const resolvedPosition = calculateCollapse(
-      piece.positions,
-      triggerPosition
-    );
-    piece.collapseTo(resolvedPosition);
+// const handleCollapse = (piece, triggerPosition) => {
+//   if (piece.isSuperposed) {
+//     const resolvedPosition = calculateCollapse(
+//       piece.positions,
+//       triggerPosition
+//     );
+//     piece.collapseTo(resolvedPosition);
 
-    if (piece.entangledWith) {
-      piece.entangledWith.collapseTo(
-        calculateEntangledCollapse(piece.entangledWith.positions)
-      );
-    }
-  }
-};
+//     if (piece.entangledWith) {
+//       piece.entangledWith.collapseTo(calculateEntangledCollapse());
+//     }
+//   }
+// };
 
-const calculateCollapse = (positions, trigger) => {
-  // Implement probability-based or deterministic collapse
-  return positions[Math.floor(Math.random() * positions.length)];
-};
-
-export { handleCollapse, calculateCollapse };
+// const calculateCollapse = (positions, trigger) => {
+//   // Implement probability-based or deterministic collapse
+//   return positions[Math.floor(Math.random() * positions.length)];
+// };
