+++ app/quantumChess/useQuantumMove.ts
@@ -1,30 +1,27 @@
 export default function useQuantumMove() {
-  const [quantumMoves, setQuantumMoves] = useState({
-    superpositions: 3,
-    entanglements: 2,
-    tunnels: 1,
-  });
-
-  const handleSuperposition = (piece, positions) => {
-    if (quantumMoves.superpositions > 0) {
-      piece.setSuperposition(positions);
-      setQuantumMoves((prev) => ({
-        ...prev,
-        superpositions: prev.superpositions - 1,
-      }));
-    }
-  };
-
-  const handleEntanglement = (piece1, piece2) => {
-    if (quantumMoves.entanglements > 0) {
-      piece1.entangleWith(piece2);
-      piece2.entangleWith(piece1);
-      setQuantumMoves((prev) => ({
-        ...prev,
-        entanglements: prev.entanglements - 1,
-      }));
-    }
-  };
-
+  // const [quantumMoves, setQuantumMoves] = useState({
+  //   superpositions: 3,
+  //   entanglements: 2,
+  //   tunnels: 1,
+  // });
+  // const handleSuperposition = (piece, positions) => {
+  //   if (quantumMoves.superpositions > 0) {
+  //     piece.setSuperposition(positions);
+  //     setQuantumMoves((prev) => ({
+  //       ...prev,
+  //       superpositions: prev.superpositions - 1,
+  //     }));
+  //   }
+  // };
+  // const handleEntanglement = (piece1, piece2) => {
+  //   if (quantumMoves.entanglements > 0) {
+  //     piece1.entangleWith(piece2);
+  //     piece2.entangleWith(piece1);
+  //     setQuantumMoves((prev) => ({
+  //       ...prev,
+  //       entanglements: prev.entanglements - 1,
+  //     }));
+  //   }
+  // };
   // ... similar handlers for tunneling
 }
