+++ app/user/initialProfileData.ts
@@ -1,0 +1,14 @@
+export const initialProfileData = {
+  currentRank: "N/A",
+  xp: 0,
+  achievementsUnlocked: 0,
+  quantumChessSkillsProgress: 0,
+  puzzleSolvingProgress: 0,
+  learningProgress: 0,
+  quantumChessWins: 0,
+  quantumChessTotalGames: 0,
+  quantumChessCurrentStreak: 0,
+  puzzlesSolved: 0,
+  puzzleAverageTimeSeconds: null,
+  puzzlePerfectSolutions: 0,
+};
